rename: page 디렉터리의 내의 루트 파일명 변경 (index.tsx로 통일)
